Move config base types from telemetry-utils to core-interfaces
