+++ azure/packages/azure-client/src/interfaces.ts
@@ -2,11 +2,13 @@
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
  */
-import { type ITelemetryBaseLogger } from "@fluidframework/core-interfaces";
+import {
+	type IConfigProviderBase,
+	type ITelemetryBaseLogger,
+} from "@fluidframework/core-interfaces";
 import { type IMember, type IServiceAudience } from "@fluidframework/fluid-static";
 import { type IUser } from "@fluidframework/protocol-definitions";
 import { type ITokenProvider } from "@fluidframework/routerlicious-driver";
-import { type IConfigProviderBase } from "@fluidframework/telemetry-utils";
 import { type ICompressionStorageConfig } from "@fluidframework/driver-utils";
 
 /**
