+++ azure/packages/test/end-to-end-tests/src/test/containerCreate.spec.ts
@@ -10,8 +10,9 @@ import { ContainerSchema } from "@fluidframework/fluid-static";
 import { SharedMap } from "@fluidframework/map";
 import { timeoutPromise } from "@fluidframework/test-utils";
 
-import { ConfigTypes, IConfigProviderBase, MockLogger } from "@fluidframework/telemetry-utils";
+import { MockLogger } from "@fluidframework/telemetry-utils";
 import { ConnectionState } from "@fluidframework/container-loader";
+import { ConfigTypes, IConfigProviderBase } from "@fluidframework/core-interfaces";
 import { createAzureClient } from "./AzureClientFactory";
 
 const configProvider = (settings: Record<string, ConfigTypes>): IConfigProviderBase => ({
