+++ packages/common/core-interfaces/src/index.ts
@@ -78,3 +78,4 @@ export type {
 } from "./logger";
 export { LogLevel } from "./logger";
 export { FluidObjectProviderKeys, FluidObject, FluidObjectKeys } from "./provider";
+export { ConfigTypes, IConfigProviderBase } from "./config";
