+++ packages/runtime/container-runtime/src/test/gc/gcUnitTestHelpers.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { ConfigTypes, IConfigProviderBase } from "@fluidframework/telemetry-utils";
+import { ConfigTypes, IConfigProviderBase } from "@fluidframework/core-interfaces";
 import { ReadAndParseBlob } from "@fluidframework/runtime-utils";
 
 export const configProvider = (settings: Record<string, ConfigTypes>): IConfigProviderBase => ({
