+++ packages/service-clients/odsp-client/src/interfaces.ts
@@ -5,7 +5,7 @@
 import { type ITelemetryBaseLogger } from "@fluidframework/common-definitions";
 import type { IMember, IServiceAudience } from "@fluidframework/fluid-static";
 import { type IUser } from "@fluidframework/protocol-definitions";
-import { type IConfigProviderBase } from "@fluidframework/telemetry-utils";
+import { IConfigProviderBase } from "@fluidframework/core-interfaces";
 import { IOdspTokenProvider } from "./token";
 
 /**
