+++ packages/test/test-end-to-end-tests/src/test/messageSize.spec.ts
@@ -17,7 +17,12 @@ import {
 } from "@fluidframework/test-utils";
 import { describeNoCompat, itExpects } from "@fluid-private/test-version-utils";
 import { IContainer } from "@fluidframework/container-definitions";
-import { FluidErrorTypes, IErrorBase } from "@fluidframework/core-interfaces";
+import {
+	ConfigTypes,
+	FluidErrorTypes,
+	IConfigProviderBase,
+	IErrorBase,
+} from "@fluidframework/core-interfaces";
 import { FlushMode } from "@fluidframework/runtime-definitions";
 import { CompressionAlgorithms, ContainerMessageType } from "@fluidframework/container-runtime";
 import {
@@ -25,8 +30,7 @@ import {
 	ISequencedDocumentMessage,
 	MessageType,
 } from "@fluidframework/protocol-definitions";
-import { ConfigTypes, GenericError, IConfigProviderBase } from "@fluidframework/telemetry-utils";
-
+import { GenericError } from "@fluidframework/telemetry-utils";
 describeNoCompat("Message size", (getTestObjectProvider) => {
 	const mapId = "mapId";
 	const registry: ChannelFactoryRegistry = [[mapId, SharedMap.getFactory()]];
