+++ packages/tools/replay-tool/src/helpers.ts
@@ -10,10 +10,14 @@ import { ILoaderOptions, Loader } from "@fluidframework/container-loader";
 import { ContainerRuntime, IContainerRuntimeOptions } from "@fluidframework/container-runtime";
 import { IDocumentServiceFactory, IResolvedUrl } from "@fluidframework/driver-definitions";
 import { IFileSnapshot } from "@fluidframework/replay-driver";
-import { ConfigTypes, IConfigProviderBase } from "@fluidframework/telemetry-utils";
 import { getNormalizedSnapshot, ISnapshotNormalizerConfig } from "@fluidframework/tool-utils";
 import stringify from "json-stable-stringify";
-import { FluidObject, ITelemetryLogger } from "@fluidframework/core-interfaces";
+import {
+	ConfigTypes,
+	FluidObject,
+	IConfigProviderBase,
+	ITelemetryLogger,
+} from "@fluidframework/core-interfaces";
 import { assert } from "@fluidframework/core-utils";
 import {
 	excludeChannelContentDdsFactories,
