+++ packages/utils/telemetry-utils/src/config.ts
@@ -2,26 +2,14 @@
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
  */
-import { ITelemetryBaseLogger } from "@fluidframework/core-interfaces";
+import {
+	ITelemetryBaseLogger,
+	IConfigProviderBase,
+	ConfigTypes,
+} from "@fluidframework/core-interfaces";
 import { Lazy } from "@fluidframework/core-utils";
 import { createChildLogger, tagCodeArtifacts } from "./logger";
 import { ITelemetryLoggerExt } from "./telemetryTypes";
-
-/**
- * Types supported by {@link IConfigProviderBase}.
- *
- * @internal
- */
-export type ConfigTypes = string | number | boolean | number[] | string[] | boolean[] | undefined;
-
-/**
- * Base interface for providing configurations to enable/disable/control features.
- *
- * @internal
- */
-export interface IConfigProviderBase {
-	getRawConfig(name: string): ConfigTypes;
-}
 
 /**
  * Explicitly typed interface for reading configurations.
