+++ packages/utils/telemetry-utils/src/index.ts
@@ -5,11 +5,9 @@
 export {
 	createChildMonitoringContext,
 	MonitoringContext,
-	IConfigProviderBase,
 	sessionStorageConfigProvider,
 	mixinMonitoringContext,
 	IConfigProvider,
-	ConfigTypes,
 	loggerToMonitoringContext,
 } from "./config";
 export {
@@ -81,3 +79,22 @@ export {
 	ITelemetryPropertiesExt,
 	TelemetryEventCategory,
 } from "./telemetryTypes";
+
+/**
+ * Types supported by {@link IConfigProviderBase}.
+ * @deprecated Use ConfigTypes from fluidFramework/core-interfaces
+ *
+ * @internal
+ */
+export type ConfigTypes = string | number | boolean | number[] | string[] | boolean[] | undefined;
+
+/**
+ * Base interface for providing configurations to enable/disable/control features.
+ *
+ * @deprecated Use IConfigProviderBase from fluidFramework/core-interfaces
+ *
+ * @internal
+ */
+export interface IConfigProviderBase {
+	getRawConfig(name: string): ConfigTypes;
+}
